steps:calories and weight

+++ mobile/src/main/java/fit/asta/health/tools/walking/view/steps_counter/StepCounterUIState.kt
@@ -1,11 +1,11 @@
 package fit.asta.health.tools.walking.view.steps_counter
 
 data class StepCounterUIState(
-    val calories: Int = 0,
+    val calories: Double = 0.0,
     val steps: Int = 0,
     val distance: Int = 0,
     val speed: Float = 0f,
     val time: Long = 0,
     val heartRate:Int=72,
-    val weightLoosed:Float=0f
+    val weightLoosed:Double=0.0
 )
